[Front] refactor(agent-builder): move skillSheet to capabilities_sheet
