+++ front/components/agent_builder/capabilities/capabilities_sheet/SkillWithRelationsDetailsSheetContent.tsx
